requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.13.3
async-timeout==4.0.3
asyncpg==0.29.0
attrs==24.2.0
et-xmlfile==1.1.0
frozenlist==1.4.1
greenlet==3.1.1
idna==3.10
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.1.0
openpyxl==3.1.5
pip-autoremove==0.10.0
prometheus-async==22.2.0
prometheus_client==0.21.0
propcache==0.2.0
pyaes==1.6.1
pyasn1==0.6.0
python-dotenv==1.0.1
rsa==4.9
SQLAlchemy==2.0.35
Telethon==1.36.0
typing_extensions==4.12.2
wrapt==1.16.0
yarl==1.15.1
